+++ tests/testthat/test-utilities.R
@@ -1,0 +1,9 @@
+## Testing utilities
+
+test_that("parent.traits works", {
+
+})
+
+test_that("snapshot works", {
+    
+})